Move check for store_artifact parameter to functions module
